refactor(product): product and category id type modified

+++ service/product/src/main/resources/db/migration/V1__init_database.sql
@@ -1,18 +1,18 @@
 create table if not exists category
 (
-    id integer not null primary key,
+    id bigint not null primary key,
     description varchar(255),
     name varchar(255)
 );
 
 create table if not exists product
 (
-    id integer not null primary key,
+    id bigint not null primary key,
     description varchar(255),
     name varchar(255),
     available_quantity double precision not null,
     price numeric(38, 2),
-    category_id integer
+    category_id bigint
         constraint fk1alfsakdmqwkl references category
 );
